% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_named_list}
\alias{as_named_list}
\title{Take objects and create a list using their names}
\usage{
as_named_list(...)
}
\arguments{
\item{...}{additional parameters}
}
\description{
This avoids retyping the name of the object as in a list
}
\keyword{internal}
