% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petro.One-package.r
\docType{data}
\name{custom_stopwords}
\alias{custom_stopwords}
\title{Default custom stop words}
\format{An object of class \code{NULL} of length 0.}
\usage{
custom_stopwords
}
\description{
This is a minimal dataset of custom stopwords.
You can supply your own stopwords by editing the file stopwords.txt under
`extdata` and then importing it.
The provided dataset is a basic way to start and eliminate
common words from the paper titles during classification.

Dataset: stopwords.rda

Source:  stopwords.txt
}
\keyword{datasets}
