% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petro.One-package.r
\docType{data}
\name{custom_synonyms}
\alias{custom_synonyms}
\title{Default synonyms}
\format{An object of class \code{data.frame} with 33 rows and 2 columns.}
\usage{
custom_synonyms
}
\description{
Synonyms dataset to prevent repetition of equivalent words
during classification. Example: 2D/2-D, cased hole/cased-hole,
deep water/deepwater, etc.
The first column is for the original word; the second column is for the
replacement or standard word.

Dataset: synonyms.rda

Source:  synonyms.txt
}
\keyword{datasets}
