% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mock_data.R
\name{generate_offline_data}
\alias{generate_offline_data}
\title{Generate data for offline testing
Mockup test data}
\usage{
generate_offline_data()
}
\description{
Generate data for offline testing
Mockup test data
}
