% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{get_papers_count}
\alias{get_papers_count}
\title{Number of paper for a given query}
\usage{
get_papers_count(url)
}
\arguments{
\item{url}{char a query URL for OnePetro}
}
\description{
Obtains the number of papers being queried by the URL
}
\examples{
\dontrun{
# Example 1
url_1 <- make_search_url(query = "static gradient survey", how = "all")
get_papers_count(url_1)
#
# Example 2
url_2 <- make_search_url(query = "vertical lift performance", how = "all")
get_papers_count(url_2)
#
# Example 3
url_3 <- make_search_url(query = "inflow performance relationship", how = "all")
get_papers_count(url_3)
}
}
