% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{get_term_document_matrix}
\alias{get_term_document_matrix}
\title{A TermDocumentMatrix corpus objects}
\usage{
get_term_document_matrix(df)
}
\arguments{
\item{df}{a dataframe with paper results}
}
\value{
a list
}
\description{
Transforms a document into a VCorpus TermDocumentMatrix object
plus additional calculated matrix, row sums and words frequency objects
}
