% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{get_top_term_papers}
\alias{get_top_term_papers}
\title{Get papers for top "N" terms}
\usage{
get_top_term_papers(papers, tdm_matrix, top_terms, terms = NULL,
  verbose = FALSE)
}
\arguments{
\item{papers}{a dataframe with papers}

\item{tdm_matrix}{a Term Document Matrix}

\item{top_terms}{top 10, or top 20, etc.}

\item{terms}{a term or vector of terms to get papers from}

\item{verbose}{set to TRUE to show progress}
}
\description{
Indicate the top terms from which we want to extract papers.
For instance, if we want the papers for the top 10 terms, we set top_terms = 10.
}
