% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petro.One-package.r
\docType{data}
\name{labels}
\alias{labels}
\title{Discipline and Subject  labels dataset}
\usage{
labels(object, ...)
}
\description{
dataset containing disciplines and subjects.
The purpose is to categorize papers based on the words in their title since
OnePetro does not supply keywords or any sort of categorization.
File:  disciplines.rda
Class: data.frame
}
