% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{papers_by_publication}
\alias{papers_by_publication}
\title{Papers by publication}
\usage{
papers_by_publication(url)
}
\arguments{
\item{url}{a OnePetro query URL}
}
\description{
Generate a summary by publications. These publications could
be World Petroleum Congress, Annual Technical
Meeting, SPE Unconventional Reservoirs Conference, etc.
}
\examples{
#
# Example
my_url <- make_search_url(query = "industrial drilling", how = "all")
papers_by_publication(my_url)
}
