% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{papers_by_publisher}
\alias{papers_by_publisher}
\title{Papers by publisher}
\usage{
papers_by_publisher(url)
}
\arguments{
\item{url}{a OnePetro query URL}
}
\description{
Generate a summary by publisher.
Know publishers: OTC, SPE, etc.
}
\examples{
#
# Example
my_url <- make_search_url(query = "shale gas", how = "all")
papers_by_publisher(my_url)
}
