% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{papers_by_type}
\alias{papers_by_type}
\title{Summary by document type}
\usage{
papers_by_type(url)
}
\arguments{
\item{url}{a OnePetro page with results}
}
\description{
Generate a summary by document type.
Types are: conference-paper, journal-paper, presentation, media, other, etc.
}
\examples{
\dontrun{
#
# Example 1
my_url <- make_search_url(query = "well test", how = "all")
papers_by_type(my_url)
}
}
