% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{papers_by_year}
\alias{papers_by_year}
\title{Papers by Year}
\usage{
papers_by_year(url)
}
\arguments{
\item{url}{a OnePetro query URL}
}
\description{
Generate a summary by the year the paper was published
}
\examples{
\dontrun{
# Example
my_url <- make_search_url(query = "production automation", how = "all")
papers_by_year(my_url)
}
}
