% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petro.One-package.r
\docType{package}
\name{petro.One-package}
\alias{petro.One-package}
\title{Text mining and statistics for OnePetro papers
petro.One}
\description{
Text mining and statistics for OnePetro papers
petro.One
}
