% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{plot_cluster_dendrogram}
\alias{plot_cluster_dendrogram}
\title{Plot a dendrogram}
\usage{
plot_cluster_dendrogram(df)
}
\arguments{
\item{df}{a dataframe with paper results}
}
\description{
Plots a clustering diagram of terms
}
