% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{plot_relationships}
\alias{plot_relationships}
\title{Plot a relationship diagram with weights}
\usage{
plot_relationships(df, ..., min.freq = 25, threshold = 0.1)
}
\arguments{
\item{df}{a dataframe with paper results}

\item{...}{additional parameters}

\item{min.freq}{minimum frequency of the words to be plotted}

\item{threshold}{correlation threshold}
}
\description{
Plots a diagram with relationships between words. The lines
that link the terms are weighted according to how often the connect together
}
\examples{
\dontrun{
my_url <- make_search_url(query = "well test",
                          dc_type = "conference-paper",
                          from_year = 2017,
                          to_year   = 2018,
                          how = "all")
df <- read_multidoc(my_url)   # create a dataframe of papers
(tf <- term_frequency(df))    # create a term frequency table
min_freq <- min(head(tf, 20)$freq)
plot_relationships(df, min.freq = min_freq, threshold = 0.075)
}
}
