% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{plot_wordcloud}
\alias{plot_wordcloud}
\title{Plot a word cloud}
\usage{
plot_wordcloud(df, ..., max.words = 200, min.freq = 50)
}
\arguments{
\item{df}{A dataframe with paper results}

\item{...}{other parameters}

\item{max.words}{the maximum words to process}

\item{min.freq}{the minimum frequency of words allowed}
}
\description{
Plots a cloud plot of words where the size of the words is
determined by their frequency
}
