% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipage.R
\name{read_multidoc}
\alias{read_multidoc}
\title{Read all OnePetro papers metadata by type of document}
\usage{
read_multidoc(my_url)
}
\arguments{
\item{my_url}{OnePetro query URL}
}
\description{
Function iterates through all found document types and extracts
papers into a common dataframe
}
