% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipage.R
\name{read_multipage}
\alias{read_multipage}
\title{Reads metadata in groups of 1000 papers}
\usage{
read_multipage(url, verbose = FALSE)
}
\arguments{
\item{url}{A OnePetro query URL}
}
\description{
This function will loop over and grab data from the OnePetro
results in groups of 1000 papers at a time.
OnePetro limits the number of papers to view to 1000 papers and the query in
this function automatically sets the start counter to read them in groups.
}
