% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{read_onepetro}
\alias{read_onepetro}
\title{Read OnePetro web page given a query URL}
\usage{
read_onepetro(url)
}
\arguments{
\item{url}{char a query URL for OnePetro}
}
\description{
Read a OnePetro webpage using a query URL. Uses xml2
function read_html
}
