% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-papers.R
\name{remove_duplicates_by}
\alias{remove_duplicates_by}
\title{Remove duplicate papers by a variable}
\usage{
remove_duplicates_by(df, by = "book_title")
}
\arguments{
\item{df}{dataframe of papers}

\item{by}{variable}
}
\description{
Duplicates are removed in a dataframe containing papers
}
