% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_by_dates}
\alias{summary_by_dates}
\title{Summary by year}
\usage{
summary_by_dates(result)
}
\arguments{
\item{result}{a OnePetro page with results}
}
\description{
Generate a summary by the year the paper was published
}
\examples{
#
# Example
my_url <- make_search_url(query = "production automation", how = "all")
result <- read_onepetro(my_url)
summary_by_dates(result)
}
