% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_by_doctype}
\alias{summary_by_doctype}
\title{Summary by document type}
\usage{
summary_by_doctype(result)
}
\arguments{
\item{result}{a OnePetro page with results}
}
\description{
Generate a summary by document type.
Types are: conference-paper, journal-paper, presentation, media, other, etc.
}
\examples{
#
# Example 1
my_url <- make_search_url(query = "well test", how = "all")
result <- read_onepetro(my_url)
summary_by_doctype(result)
}
