% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{term_frequency_n_grams}
\alias{term_frequency_n_grams}
\title{Find the frequency for two or more words together}
\usage{
term_frequency_n_grams(df, gram.min = 2, gram.max = 2)
}
\arguments{
\item{df}{a dataframe with paper results}

\item{gram.min}{minimum amount of words together}

\item{gram.max}{maximum amount of words together}
}
\description{
Use this function when trying to find frequency of two or more words
}
