% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_example}
\alias{use_example}
\title{Unpack an example}
\usage{
use_example(which_one = NULL)
}
\arguments{
\item{which_one}{example number to use}
}
\description{
Examples are zipped to save disk space and prevent R complaining
while creating the package
}
