% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{term_frequency}
\alias{term_frequency}
\title{Word Frequency Dataframe}
\usage{
term_frequency(df, gram.min = 1, gram.max = 1)
}
\arguments{
\item{df}{a dataframe with paper results}
}
\description{
Returns a dataframe of words vs frequency
}
